flytt internstatistikk-kode fra fagsak-context og inn i egen hook
